Fix SBCL internal variable use.
The code for deferring warnings on SBCL uses the internal variable
`sb-c::*aborted-compilation-unit-count*` whose behavior has changed.
Now we catch errors attempting to access these and other SB-C internals.

This solution may not be entirely satisfactory, since it's not entirely
clear to me what this variable is for in SBCL.  It appears only to be
used to assemble messages in UIOP, so I believe this should be
acceptable.